[REF] frutilight: Several changes in sales
This PR delivers the following tasks:
- 32179: [ADD] frutilight: Add default customer
- 32180: [REF] frutilight: Enable shipping address, set domain and label
- 32181: [ADD] frutilight: Add fields to split sale order's customer
  reference
- 32182: [REF] frutilight: Hide sale order's customer reference fields
- 32186: [ADD] frutilight: Add field for delivery status, change labels
  and more
- 32183: [ADD] frutilight: Add automated action to set consignee on sale
  orders
- 32189: [ADD] frutilight: Add models to store sample indicators
- 32188: [REF] frutilight: Group sale orders by date

Below are all commit messages.

[ADD] frutilight: Add default customer

Since the customer used on sales will always be the same, a new customer
is created on data and the default customers for sale orders is set to
that customer.

[REF] frutilight: Enable shipping address, set domain and label

This commits performs the following changes:
- Enable option for customer addresses, so shipping addresses may be
  used on sale orders
- Allow to choose only customers when selecting shipping addresses
- Set label to "consignee" to the invoicing address field
- Put Shipping address before invoicing address
- By default, Fill invoicing address with the invoice contact of the
  shipping address, and allow to select only invoice contacts
- Show the address on the shipping address and invoice address fields,
  as done in the customer.

[ADD] frutilight: Add fields to split sale order's customer reference

This commit adds two fields to sale orders, to make possible to fill its
prefix and suffix separatelly.

[REF] frutilight: Hide sale order's customer reference fields

Since there are two new fields to fill the customer reference (prefix
and code), there is no need for showing them all at the same time.

For that reason, the custom fields are shown only when the sale order is
in draft mode, while the native field is shown only when confirmed.

[ADD] frutilight: Add field for delivery status, change labels and more

This commit performs the following changes:
- Add a field to sale orders to store the delivery status: undelivered,
  partially delivered or fully delivered
- Change lavels of states to "draft" and "confirmed", and hide remaining
  status
- Hide buttons to print or send by e-mail
- Add menus and submenus to filter sale orders by state, delivery status
and invoice status

[ADD] frutilight: Add automated action to set consignee on sale orders

When the customer is changed, the consignee (shipping address) is
cleared.

[ADD] frutilight: Add models to store sample indicators

Add all required resources to store and handle sample indicators:
models, views, data and ACLs.

[REF] frutilight: Group sale orders by date

Now all sale orders, both quotations and confirmed ones, are grouped by
default by their date order, according to their week.