[FIX] frutilight: wrong amount currency in fruit reception
Odoo made a change recently to force that any account move line
that came from a stock move, always use as data stock move's
confirmation date.

https://github.com/odoo/odoo/commit/fda2f3122b433e04687fe3282dcc8f2a34cd9056

In Frutilight context it isn't functional that behaviour because
the amount currency for Fruit Reception always depends on Unload
Date.

This fix was made to change the behaviour to Frutilight's business
model.

Ticket #24834